Cleaned up syncservice
Extension methods for DI.
Controller response types.
Removed ghost project.
Removed IIS.
