+++ LOUPE_Backend/SynchronizationService.API/appsettings.Development.json
