+++ LOUPE_Backend/SynchronizationService.API/appsettings.json
