--- LOUPE_Backend/SynchronizationService/SynchronizationService.API.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SynchronizationService.Core.API\SynchronizationService.Core.API.csproj" />
-    <ProjectReference Include="..\SynchronizationService.DataLayer\SynchronizationService.DataLayer.csproj" />
-  </ItemGroup>
-
-</Project>
